# aspnetcore-server/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Room Reservation API",
    "description" : "A Simple IP Address API",
    "version" : "0.1.9"
  },
  "servers" : [ {
    "url" : "http://localhost:5000/api"
  } ],
  "paths" : {
...
    "/cities" : {
      "get" : {
        "tags" : [ "city" ],
        "summary" : "Returns a list of cities",
        "responses" : {
          "200" : {
            "description" : "A json array of cities",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/City"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "404" : {
            "description" : "Cities not found"
          }
        }
      },
      "post" : {
        "tags" : [ "city" ],
        "summary" : "Creates a new city",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/City"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/City"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict"
          }
        }
      }
    },
    "/cities/{cityId}" : {
      "get" : {
        "tags" : [ "city" ],
        "summary" : "Returns a city by ID",
        "parameters" : [ {
          "name" : "cityId",
          "in" : "path",
          "description" : "ID of a city",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/City"
                }
              }
            }
          },
          "404" : {
            "description" : "City not found"
          }
        }
      },
      "put" : {
        "tags" : [ "city" ],
        "summary" : "Update an existing room",
        "parameters" : [ {
          "name" : "cityId",
          "in" : "path",
          "description" : "ID of a city",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/City"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Updated"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "City not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "city" ],
        "summary" : "Deletes a city",
        "parameters" : [ {
          "name" : "cityId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "City not found"
          }
        }
      }
    },
    "/jobs" : {
      "get" : {
        "tags" : [ "job" ],
        "summary" : "Returns a list of jobs",
        "responses" : {
          "200" : {
            "description" : "An array of jobs",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Job"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "404" : {
            "description" : "Jobs not found"
          }
        }
      },
      "post" : {
        "tags" : [ "job" ],
        "summary" : "Creates a new job",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Job"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Job"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict"
          }
        }
      }
    },
    "/jobs/{jobId}" : {
      "get" : {
        "tags" : [ "job" ],
        "summary" : "Returns a job by ID",
        "parameters" : [ {
          "name" : "jobId",
          "in" : "path",
          "description" : "ID of a job",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Job"
                }
              }
            }
          },
          "404" : {
            "description" : "Job not found"
          }
        }
      },
      "put" : {
        "tags" : [ "job" ],
        "summary" : "Update an existing job",
        "parameters" : [ {
          "name" : "jobId",
          "in" : "path",
          "description" : "ID of a job",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Job"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad request"
          },
          "404" : {
            "description" : "Job not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "job" ],
        "summary" : "Deletes a job",
        "parameters" : [ {
          "name" : "jobId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "Job not found"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "City" : {
        "type" : "object",
        "properties" : {
          "cityId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "cityName" : {
            "type" : "string"
          }
        },
        "example" : {
          "cityName" : "cityName",
          "cityId" : 1
        }
      },
      "Job" : {
        "type" : "object",
        "properties" : {
          "jobId" : {
            "type" : "integer"
          },
          "jobTitle" : {
            "type" : "string"
          }
        },
        "example" : {
          "jobId" : 3,
          "jobTitle" : "jobTitle"
        }
      },
...
    }
  }
}